+++ cmd/function_undeploy.go
@@ -4,6 +4,7 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/vertigobr/safira/pkg/utils"
 	"os"
 
 	"github.com/spf13/cobra"
@@ -12,9 +13,9 @@ import (
 )
 
 var functionUndeployCmd = &cobra.Command{
-	Use:     "undeploy [FUNCTION_NAME]",
-	Short:   "Remove a function from the cluster",
-	Long:    "Remove a function from the cluster",
+	Use:   "undeploy [FUNCTION_NAME]",
+	Short: "Remove a function from the cluster",
+	Long:  "Remove a function from the cluster",
 	Example: `To remove the function from a project, run:
 
     $ safira function undeploy function-name
@@ -22,14 +23,15 @@ var functionUndeployCmd = &cobra.Command{
 or if you want to remove all functions from a project, execute:
 
     $ safira function undeploy -A`,
-	PreRunE: preRunFunctionUndeploy,
-	RunE:    runFunctionUndeploy,
+	PreRunE:                    preRunFunctionUndeploy,
+	RunE:                       runFunctionUndeploy,
 	SuggestionsMinimumDistance: 1,
 }
 
 func init() {
 	functionCmd.AddCommand(functionUndeployCmd)
-	functionUndeployCmd.Flags().BoolP("all-functions", "A", false, "deploy all functions")
+	functionUndeployCmd.Flags().BoolP("all-functions", "A", false, "undeploy all functions")
+	functionUndeployCmd.Flags().Bool("remove-swagger", false, "undeploy swagger ui")
 	functionUndeployCmd.Flags().String("kubeconfig", kubeconfigPath, "set kubeconfig to remove function")
 	functionUndeployCmd.Flags().StringP("namespace", "n", functionsNamespace, "set namespace to undeploy")
 }
@@ -49,6 +51,7 @@ func runFunctionUndeploy(cmd *cobra.Command, args []string) error {
 	all, _ := cmd.Flags().GetBool("all-functions")
 	kubeconfigFlag, _ := cmd.Flags().GetString("kubeconfig")
 	namespaceFlag, _ := cmd.Flags().GetString("namespace")
+	removeSwaggerUiFlag, _ := cmd.Flags().GetBool("remove-swagger")
 
 	functions, err := stack.GetAllFunctions()
 	if err != nil {
@@ -57,12 +60,12 @@ func runFunctionUndeploy(cmd *cobra.Command, args []string) error {
 
 	if all {
 		for index, _ := range functions {
-			if err := removeDeploy(index, namespaceFlag, kubeconfigFlag, "Function", verboseFlag); err != nil {
+			if err := removeFunction(index, namespaceFlag, kubeconfigFlag, verboseFlag); err != nil {
 				return err
 			}
 
-			if swaggerFile := checkSwaggerFileExist(functions[index].Handler); len(swaggerFile) > 1 {
-				if err := removeDeploy(index + "-swagger-ui", kubeconfigFlag, "default", "Swagger", verboseFlag); err != nil {
+			for plugin, _ := range functions[index].Plugins {
+				if err := removePlugin(fmt.Sprintf("%s-%s", index, plugin), kubeconfigFlag, verboseFlag); err != nil {
 					return err
 				}
 			}
@@ -70,12 +73,12 @@ func runFunctionUndeploy(cmd *cobra.Command, args []string) error {
 	} else {
 		for index, functionArg := range args {
 			if checkFunctionExists(args[index], functions) {
-				if err := removeDeploy(functionArg, namespaceFlag, kubeconfigFlag, "Function", verboseFlag); err != nil {
+				if err := removeFunction(functionArg, namespaceFlag, kubeconfigFlag, verboseFlag); err != nil {
 					return err
 				}
 
-				if swaggerFile := checkSwaggerFileExist(functions[functionArg].Handler); len(swaggerFile) > 1 {
-					if err := removeDeploy(functionArg + "-swagger-ui", kubeconfigFlag, "default", "Swagger", verboseFlag); err != nil {
+				for plugin, _ := range functions[functionArg].Plugins {
+					if err := removePlugin(fmt.Sprintf("%s-%s", functionArg, plugin), kubeconfigFlag, verboseFlag); err != nil {
 						return err
 					}
 				}
@@ -85,24 +88,64 @@ func runFunctionUndeploy(cmd *cobra.Command, args []string) error {
 		}
 	}
 
+	if removeSwaggerUiFlag {
+		repoName, err := utils.GetCurrentFolder()
+		if err != nil {
+			return err
+		}
+
+		if err := removeSwaggerUi(fmt.Sprintf("%s-swagger-ui", repoName), kubeconfigFlag, verboseFlag); err != nil {
+			return err
+		}
+	}
+
 	return nil
 }
 
-func removeDeploy(name, namespace, kubeconfigFlag, title string, verboseFlag bool) error {
+func removeFunction(name, namespace, kubeconfigFlag string, verboseFlag bool) error {
 	k8sClient, err := k8s.GetClient(kubeconfigFlag)
 	if err != nil {
 		return fmt.Errorf("cluster não encontrado!\n")
 	}
 
-	if err := k8s.RemoveDeployment(k8sClient, name, namespace, title, verboseFlag); err != nil {
+	if err := k8s.RemoveFunction(name, namespace, "Function", kubeconfigFlag, verboseFlag); err != nil {
 		return err
 	}
 
-	if err := k8s.RemoveService(k8sClient, name, namespace, title, verboseFlag); err != nil {
+	if err := k8s.RemoveService(k8sClient, name, "default", "Function", verboseFlag); err != nil {
 		return err
 	}
 
-	if err := k8s.RemoveIngress(k8sClient, name, namespace, title, verboseFlag); err != nil {
+	if err := k8s.RemoveIngress(k8sClient, name, "default", "Function", verboseFlag); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func removeSwaggerUi(name, kubeconfigFlag string, verboseFlag bool) error {
+	k8sClient, err := k8s.GetClient(kubeconfigFlag)
+	if err != nil {
+		return fmt.Errorf("cluster não encontrado!\n")
+	}
+
+	if err := k8s.RemoveDeployment(k8sClient, name, "default", "Swagger UI", verboseFlag); err != nil {
+		return err
+	}
+
+	if err := k8s.RemoveService(k8sClient, name, "default", "Swagger UI", verboseFlag); err != nil {
+		return err
+	}
+
+	if err := k8s.RemoveIngress(k8sClient, name, "default", "Swagger UI", verboseFlag); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func removePlugin(name, kubeconfigFlag string, verboseFlag bool) error {
+	if err := k8s.RemovePlugin(name, kubeconfigFlag, verboseFlag); err != nil {
 		return err
 	}